DDA-259: Move lenses to sidebar
